refactor: Extract UI components to AppUIScreen.kt
This commit moves UI-related composables from `MainActivity.kt` to a new file `AppUIScreen.kt` and introduces `AppUIScreenErrorLogic.kt` which includes enhanced error handling and loading states.

- **MainActivity.kt:**
    - Simplified to primarily set up the `RokuAppTheme` and `Scaffold`.
    - Now calls `AppUIScreen` composable to render the main UI.
    - Removed `AppScreen`, `AppList`, and `AppItem` composables as they are moved.
    - Updated `AppScreenPreview` to use `AppUIScreen`.

- **AppUIScreen.kt (New File):**
    - Contains the `AppUIScreen`, `AppList`, and `AppItem` composables, previously in `MainActivity.kt`.
    - `AppUIScreen` observes the `AppViewModel` state and displays either an error message with a retry button or the list of apps.
    - `AppList` displays a `LazyColumn` of `AppItem`s.
    - `AppItem` displays individual app details in a `Card`.

- **AppUIScreenErrorLogic.kt (New File):**
    - Introduced `AppUIScreenErrorLogic`, `AppList1`, and `AppItem1` composables.
    - `AppUIScreenErrorLogic` enhances the UI state handling by:
        - Displaying a `CircularProgressIndicator` and "Loading apps..." text when `state.isLoading` is true.
        - Displaying an error message and a "Retry" button when `state.error` is not null.
        - Displaying "No apps found" when `state.apps` is empty and not loading or in an error state.
        - Displaying the list of apps using `AppList1` otherwise.
    - `AppList1` and `AppItem1` are similar in structure to `AppList` and `AppItem` respectively.
